backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
python-multipart==0.0.9
google-genai==0.3.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic[email]
